--- a/.travis.yml
+++ b/.travis.yml
@@ -1,8 +1,0 @@
-language: go
-
-install: go get -t ./...
-
-go:
-  - 1.2
-  - 1.3
-  - tip