--- a/circle.yml
+++ b/circle.yml
@@ -1,0 +1,26 @@
+machine:
+  services:
+    - docker
+  environment:
+    PKG: ${CIRCLE_PROJECT_REPONAME}
+    URL: github.com/${CIRCLE_PROJECT_USERNAME}/${PKG}
+    DOCKER: >
+      docker run -i -t
+      -v ${GOPATH%%:*}:/go
+      -v ${HOME}/${PKG}:/go/src/${URL}
+      -v /var/run/docker.sock:/run/docker.sock
+      -v $(which docker):/usr/bin/docker:ro
+      -v $(which docker-compose):/usr/bin/docker-compose:ro
+      -w /go/src/${URL}
+
+dependencies:
+  pre:
+    - docker pull segment/golang:latest
+  override:
+    - ${DOCKER} -v ${HOME}/.ssh:/root/.ssh:ro segment/golang:latest get
+
+test:
+  override:
+    - ${DOCKER} segment/golang:latest vet lint test bench
+  post:
+    - ${DOCKER} segment/golang:latest build