@segmentio/infra 
@vinceprignano 

I came across this repo as I was digging the dependencies, it was advertised as a fast implementation of a function that converts a string to underscore notation... didn't seem really optimized tho :P

Before:
```
$ go test -bench .
PASS
BenchmarkSnakecase-4	 1000000	      1285 ns/op
ok  	github.com/segmentio/go-snakecase	1.307s
```

After:
```
$ go test -bench .
PASS
BenchmarkSnakecase-4	 5000000	       235 ns/op
ok  	github.com/segmentio/go-snakecase	1.451s
```

PS: I also added a circle.yml file using the docker builder and a .gitignore file.